+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 ## [unreleased]
 
+ * Simple Security Provider: Allow to set *extraData* for users
  * Fixed broken support for older browsers (IE11)
 
 ## 1.5.3 (October 18, 2020)